+++ src/BigDecimal.php
@@ -13,10 +13,8 @@ use Override;
 
 /**
  * Immutable, arbitrary-precision signed decimal numbers.
- *
- * @psalm-immutable
  */
-final class BigDecimal extends BigNumber
+final readonly class BigDecimal extends BigNumber
 {
     /**
      * The unscaled value of this decimal number.
@@ -25,14 +23,14 @@ final class BigDecimal extends BigNumber
      * No leading zero must be present.
      * No leading minus sign must be present if the value is 0.
      */
-    private readonly string $value;
+    private string $value;
 
     /**
      * The scale (number of digits after the decimal point) of this decimal number.
      *
      * This must be zero or more.
      */
-    private readonly int $scale;
+    private int $scale;
 
     /**
      * Protected constructor. Use a factory method to obtain an instance.